Export prisma client from separate file to avoid circular dependencies
