+++ typescript/db/src/resolvers/__tests-skipped-while-not-handled-on-CI__/membership.ts
@@ -8,7 +8,7 @@ import {
   MutationCreateWorkspaceArgs,
   Workspace,
 } from "@labelflow/graphql-types";
-import { prisma } from "../../repository";
+import { prisma } from "../../repository/prisma-client";
 import { client, user } from "../../dev/apollo-client";
 
 // @ts-ignore
