+++ typescript/db/src/resolvers/__tests-skipped-while-not-handled-on-CI__/workspace.ts
@@ -9,7 +9,7 @@ import {
   Workspace,
   WorkspaceType,
 } from "@labelflow/graphql-types";
-import { prisma } from "../../repository";
+import { prisma } from "../../repository/prisma-client";
 import { client, user } from "../../dev/apollo-client";
 import { WorkspacePlan } from ".prisma/client";
